+++ backend/db.js
@@ -1,11 +1,25 @@
 const { Pool } = require('pg');
 
-// Use the DATABASE_URL environment variable for PostgreSQL connection
-const pool = new Pool({
+// Database configuration
+const dbConfig = {
   connectionString: process.env.DATABASE_URL,
-  ssl: {
-    rejectUnauthorized: false, // Required
-  },
+  // SSL configuration for production (Render) vs development (local)
+  ssl: process.env.NODE_ENV === 'production' 
+    ? { rejectUnauthorized: false } 
+    : false
+};
+
+// Create connection pool
+const pool = new Pool(dbConfig);
+
+// Test connection
+pool.on('connect', () => {
+  console.log('Connected to PostgreSQL database');
+});
+
+pool.on('error', (err) => {
+  console.error('Unexpected error on idle client', err);
+  process.exit(-1);
 });
 
 module.exports = pool;
