+++ backend/middleware/errorHandler.js
@@ -1,0 +1,34 @@
+/**
+ * Global error handling middleware
+ */
+const errorHandler = (err, req, res, next) => {
+  console.error('Error:', err);
+
+  // Default error response
+  let statusCode = 500;
+  let message = 'Internal Server Error';
+
+  // Handle specific error types
+  if (err.name === 'ValidationError') {
+    statusCode = 400;
+    message = err.message;
+  } else if (err.name === 'UnauthorizedError') {
+    statusCode = 401;
+    message = 'Unauthorized';
+  } else if (err.code === 'ECONNREFUSED') {
+    statusCode = 503;
+    message = 'Database connection failed';
+  } else if (err.code === 'ENOTFOUND') {
+    statusCode = 503;
+    message = 'External service unavailable';
+  } else if (err.message) {
+    message = err.message;
+  }
+
+  res.status(statusCode).json({
+    error: message,
+    ...(process.env.NODE_ENV === 'development' && { stack: err.stack })
+  });
+};
+
+module.exports = errorHandler;
