+++ backend/package.json
@@ -1,19 +1,28 @@
 {
-  "name": "backend",
+  "name": "asteroid-backend",
   "version": "1.0.0",
-  "main": "index.js",
+  "description": "Backend API for asteroid data",
+  "main": "app.js",
   "scripts": {
+    "start": "node app.js",
+    "dev": "nodemon app.js",
+    "init-db": "node scripts/initDatabase.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "description": "",
   "dependencies": {
-    "axios": "^1.10.0",
+    "express": "^4.18.2",
     "cors": "^2.8.5",
-    "dotenv": "^16.5.0",
-    "express": "^5.1.0",
-    "pg": "^8.16.0"
-  }
+    "axios": "^1.6.0",
+    "pg": "^8.11.0",
+    "dotenv": "^16.3.1"
+  },
+  "devDependencies": {
+    "nodemon": "^3.0.1"
+  },
+  "engines": {
+    "node": ">=16.0.0"
+  },
+  "keywords": ["asteroid", "nasa", "api"],
+  "author": "",
+  "license": "ISC"
 }
